# internal/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garbage Truck Go core",
    "version": "0.2.0",
    "description": "Local synthetic prototype. Mutations require the gt_csrf cookie and a matching X-CSRF-Token."
  },
  "components": {
    "securitySchemes": {
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
    "/api/gaps": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
    "/api/fields": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
    "/api/candidates": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      }
    },
    "/api/runs": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        }
      },
      "post": {
        "responses": {
          "202": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "security": [
          {
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "preview",
                      "review"
                    ],
                    "default": "preview"
                  },
                  "fields": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 1,
                    "items": {
                      "const": "country"
                    },
                    "default": [
                      "country"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/runs/{id}/{action}": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "pause",
                "resume",
                "cancel"
              ]
            }
          }
        ],
        "security": [
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/runs/{id}/suggestions": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/runs/{id}/writebacks": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/suggestions/{id}/approve": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "value": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/suggestions/{id}/reject": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/writebacks/{id}/undo": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/demo/candidates/{id}/country": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "422": {
            "description": "Invalid input"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "country": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}